pyproject.toml
[project]
name = "yandex-gpt-tutorial"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "openai>=1.10.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "mypy",
    "black",
    "ruff",
]
